相关性变更

+++ src/main/java/com/connor/hozon/bom/bomSystem/iservice/cfg/vwo/IHzVWOManagerService.java
@@ -276,5 +276,5 @@ public interface IHzVWOManagerService {
 
     void doQueryCmcrDetailChangBefor2(Map<String,Object> map, Long vwoId);
 
-    Map<String,Object> getRelevance(Integer orderChangeId, String projectUid);
+    Map<String,Object> getRelevance(Long orderChangeId, String projectUid);
 }
